moved donation types to own controller
